feat: hydrate url for database stored files with a data url
