+++ storage/database_storer.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/teamkeel/keel/db"
+	"github.com/vincent-petithory/dataurl"
 )
 
 type DbStore struct {
@@ -15,6 +16,12 @@ type DbStore struct {
 var _ Storer = &DbStore{}
 
 const dbTable = "keel_storage" // table where files will be stored
+
+type fileData struct {
+	Filename    string
+	ContentType string
+	Data        []byte
+}
 
 // NewDbStore will return a Storage service for files that is db based
 func NewDbStore(ctx context.Context, db db.Database) (*DbStore, error) {
@@ -44,8 +51,27 @@ func (s *DbStore) setupDB(ctx context.Context) error {
 	return nil
 }
 
+// decodeDataURL will take a dataURL and return it as a FileData struct
+func decodeDataURL(dataURL string) (fileData, error) {
+	durl, err := dataurl.DecodeString(dataURL)
+	if err != nil {
+		return fileData{}, fmt.Errorf("decoding data url: %w", err)
+	}
+
+	return fileData{
+		ContentType: durl.ContentType(),
+		Filename:    durl.Params["name"],
+		Data:        durl.Data,
+	}, nil
+}
+
+func encodeDataURL(data fileData) string {
+	durl := dataurl.New(data.Data, data.ContentType, "name", data.Filename)
+	return durl.String()
+}
+
 func (s *DbStore) Store(url string) (FileInfo, error) {
-	fd, err := DecodeDataURL(url)
+	fd, err := decodeDataURL(url)
 	if err != nil {
 		return FileInfo{}, fmt.Errorf("decoding data URL: %w", err)
 	}
@@ -85,10 +111,25 @@ func (s *DbStore) GetFileInfo(key string) (FileInfo, error) {
 }
 
 func (s *DbStore) HydrateFileInfo(fi *FileInfo) (FileInfo, error) {
-	hydrated, err := s.GetFileInfo(fi.Key)
-	if err != nil {
-		return FileInfo{}, fmt.Errorf("hydrating file info: %w", err)
-	}
+	sql := `SELECT
+			filename,
+			content_type,
+			data
+		FROM ` + dbTable + ` WHERE id = ?`
 
-	return hydrated, nil
+	var fd fileData
+
+	db := s.db.GetDB().Raw(sql, fi.Key).Scan(&fd)
+	if db.Error != nil {
+		return FileInfo{}, fmt.Errorf("retrieving file data: %w", db.Error)
+	}
+	dataURL := encodeDataURL(fd)
+
+	return FileInfo{
+		Key:         fi.Key,
+		Filename:    fi.Filename,
+		ContentType: fi.ContentType,
+		Size:        fi.Size,
+		URL:         &dataURL,
+	}, nil
 }
