+++ storage/storage.go
@@ -3,8 +3,6 @@ package storage
 import (
 	"encoding/json"
 	"fmt"
-
-	"github.com/vincent-petithory/dataurl"
 )
 
 // Storer represents the interface for a file storing service that is used by the Keel runtime
@@ -36,11 +34,6 @@ type FileInfo struct {
 	URL         *string `json:"url,omitempty"`
 	Public      bool    `json:"public"`
 }
-type fileData struct {
-	Filename    string
-	ContentType string
-	Data        []byte
-}
 
 func (fi *FileInfo) ToJSON() (string, error) {
 	json, err := json.Marshal(fi)
@@ -48,18 +41,4 @@ func (fi *FileInfo) ToJSON() (string, error) {
 		return "", fmt.Errorf("marshalling to json: %w", err)
 	}
 	return string(json), nil
-}
-
-// DecodeDataURL will take a dataURL and return it as a FileData struct
-func DecodeDataURL(dataURL string) (fileData, error) {
-	durl, err := dataurl.DecodeString(dataURL)
-	if err != nil {
-		return fileData{}, fmt.Errorf("decoding data url: %w", err)
-	}
-
-	return fileData{
-		ContentType: durl.ContentType(),
-		Filename:    durl.Params["name"],
-		Data:        durl.Data,
-	}, nil
 }
